Add files via upload
Display the main page where we  click the button on add new outline and open a simple form. Then we fill the form and then click to upload our new outline upload or database and also fetch the record from the database and show the output in form of table and then in table we have three button where we edit the outline, delete the outline and also update the outline.